requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backend==0.2.4.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
drf-spectacular==0.26.4
idna==3.4
inflection==0.5.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kombu==5.3.1
oauthlib==3.2.2
prompt-toolkit==3.0.39
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2018.5
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
